requirements.txt
analytics-python==1.4.0
appdirs==1.4.4
asgiref==3.4.1
attr==0.3.1
attrs==21.2.0
azure-core==1.18.0
azure-storage-blob==12.8.1
backcall==0.2.0
backoff==1.10.0
bcrypt==3.2.0
boto==2.49.0
boto3==1.18.14
botocore==1.21.44
boxing==0.1.4
bravado==11.0.3
bravado-core==5.17.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
Django==3.1.13
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-redis-cache==3.0.0
django-rest-swagger==2.2.0
django-rq==2.3.2
django-user-agents==0.4.0
djangorestframework==3.12.4
drf-dynamic-fields==0.3.0
drf-generators==0.3.0
drf-yasg==1.20.0
entrypoints==0.3
fastapi==0.68.1
ffmpy==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.24
google-api-core==1.31.2
google-auth==1.35.0
google-cloud-appengine-logging==0.1.4
google-cloud-audit-log==0.1.1
google-cloud-core==1.4.1
google-cloud-logging==2.6.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.53.0
gradio==2.3.4
grpcio==1.40.0
h11==0.12.0
htmlmin==0.1.12
httptools==0.2.0
idna==3.2
inflection==0.5.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
kiwisolver==1.3.2
label-studio==1.3.post0
label-studio-converter==0.0.32
-e git+https://github.com/heartexlabs/label-studio-ml-backend@d009412e99d5818fb17ba4f03ba41fd3c26be7c7#egg=label_studio_ml
lockfile==0.12.2
lxml==4.6.3
markdown2==2.4.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
monotonic==1.6
msgpack==1.0.2
msrest==0.6.21
neptune-client==0.10.10
nest-asyncio==1.5.1
nltk==3.5
numpy==1.21.2
oauthlib==3.1.1
openapi-codec==1.3.2
opencv-python==4.5.3.56
ordered-set==4.0.2
packaging==21.0
pandas==1.3.3
paramiko==2.7.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.18.0
psutil==5.8.0
psycopg2-binary==2.8.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
Pygments==2.10.0
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2019.3
PyYAML==5.4.1
pyzmq==22.3.0
ray==1.6.0
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rq==1.7.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
rules==2.2
s3transfer==0.5.0
scikit-learn==0.24.1
scipy==1.7.1
sentry-sdk==1.3.1
simplejson==3.17.5
six==1.16.0
smmap==4.0.0
sqlparse==0.4.2
starlette==0.14.2
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tabulate==0.8.9
tensorboardX==2.4
threadpoolctl==2.2.0
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
ua-parser==0.10.0
ujson==4.1.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
webcolors==1.11.1
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.1
xmljson==0.2.0

mkdocs-material
mkdocs-macros-plugin
mkdocstrings

rich
